refactor: adjust path names
