+++ packages/x-adapter-platform/src/endpoint-adapters/related-tags.endpoint-adapter.ts
@@ -1,8 +1,8 @@
 import { endpointAdapterFactory, EndpointAdapterOptions } from '@empathyco/x-adapter-next';
 import { BaseRequest } from '../types/request.types';
 import { RelatedTagsResponse } from '../types/response.types';
-import { relatedTagsResponseMapper } from '../mappers/response/related-tags-response.mapper';
-import { baseRequestMapper } from '../mappers/request/base-request.mapper';
+import { relatedTagsResponseMapper } from '../mappers/responses/related-tags-response.mapper';
+import { baseRequestMapper } from '../mappers/requests/base-request.mapper';
 
 export const relatedTagsEndpointAdapterOptions: EndpointAdapterOptions<
   BaseRequest,
