+++ packages/x-adapter-platform/src/endpoint-adapters/sku-search.endpoint-adapter.ts
@@ -1,8 +1,8 @@
 import { endpointAdapterFactory, EndpointAdapterOptions } from '@empathyco/x-adapter-next';
 import { BaseRequest } from '../types/request.types';
 import { SkuSearchResponse } from '../types/response.types';
-import { skuSearchResponseMapper } from '../mappers/response/sku-search-response.mapper';
-import { baseRequestMapper } from '../mappers/request/base-request.mapper';
+import { skuSearchResponseMapper } from '../mappers/responses/sku-search-response.mapper';
+import { baseRequestMapper } from '../mappers/requests/base-request.mapper';
 
 const skuSearchEndpointAdapterOptions: EndpointAdapterOptions<BaseRequest, SkuSearchResponse> = {
   endpoint: 'https://api.{env(.)}empathy.co/search/v1/query/{instance}/skusearch',
