+++ packages/x-adapter-platform/src/mappers/__tests__/request/search-request.mapper.spec.ts
@@ -1,6 +1,6 @@
 import { Dictionary } from '@empathyco/x-utils';
 import { Filter, HierarchicalFilter, NumberRangeFilter, SimpleFilter } from '@empathyco/x-types';
-import { searchRequestMapper } from '../../request/search-request.mapper';
+import { searchRequestMapper } from '../../requests/search-request.mapper';
 import { PlatformSearchRequest, SearchRequest } from '../../../types/request.types';
 
 const filters: Dictionary<Filter[]> = {
