+++ packages/x-adapter-platform/src/mappers/requests/__tests__/next-queries-request.mapper.spec.ts
