+++ packages/x-adapter-platform/src/mappers/requests/search-request.mapper.ts
