+++ packages/x-adapter-platform/src/mappers/responses/empathize-response.mapper.ts
