+++ packages/x-adapter-platform/src/mappers/responses/next-queries-response.mapper.ts
@@ -1,7 +1,7 @@
 import { schemaMapperFactory } from '@empathyco/x-adapter-next';
 import { NextQueriesResponse } from '@empathyco/x-types';
-import { nextQueriesResponseMutableSchema } from '../../schemas/response';
-import { PlatformNextQueriesResponse } from '../../types/response/next-queries-response.model';
+import { nextQueriesResponseMutableSchema } from '../../schemas/responses';
+import { PlatformNextQueriesResponse } from '../../types/responses/next-queries-response.model';
 
 export const nextQueriesResponseMapper = schemaMapperFactory<
   PlatformNextQueriesResponse,
