+++ packages/x-adapter-platform/src/mappers/responses/search-response.mapper.ts
