+++ packages/x-adapter-platform/src/mappers/responses/top-clicked-response.mapper.ts
