+++ packages/x-adapter-platform/src/schemas/requests/next-queries-request.schema.ts
@@ -1,6 +1,6 @@
 import { createMutableSchema, Schema } from '@empathyco/x-adapter-next';
 import { NextQueriesRequest } from '@empathyco/x-types';
-import { PlatformNextQueriesRequest } from '../../types/request/next-queries-request.model';
+import { PlatformNextQueriesRequest } from '../../types/requests/next-queries-request.model';
 
 export const nextQueriesRequestMutableSchema = createMutableSchema<
   Schema<NextQueriesRequest, PlatformNextQueriesRequest>
