+++ packages/x-adapter-platform/src/schemas/responses/index.ts
