+++ packages/x-adapter-platform/src/schemas/responses/models/index.ts
