+++ packages/x-adapter-platform/src/schemas/responses/next-queries-response.schema.ts
@@ -1,6 +1,6 @@
 import { Schema, createMutableSchema } from '@empathyco/x-adapter-next';
 import { NextQueriesResponse } from '@empathyco/x-types';
-import { PlatformNextQueriesResponse } from '../../types/response/next-queries-response.model';
+import { PlatformNextQueriesResponse } from '../../types/responses/next-queries-response.model';
 import { nextQueryMutableSchema } from './models/next-query.schema';
 
 export const nextQueriesResponseMutableSchema = createMutableSchema<
