+++ packages/x-adapter-platform/src/types/requests/next-queries-request.model.ts
