+++ packages/x-adapter-platform/src/types/responses/models/next-query.model.ts
