+++ packages/x-adapter-platform/src/types/responses/next-queries-response.model.ts
